requirements.txt
Flask==2.0.1
GitPython==3.1.24